+++ hrx-delivery-woo/classes/Cronjob.php
@@ -7,7 +7,8 @@ if ( ! defined('ABSPATH') ) {
 }
 
 use HrxDeliveryWoo\Core;
-use HrxDeliveryWoo\Terminal;
+use HrxDeliveryWoo\Warehouse;
+use HrxDeliveryWoo\LocationsDelivery;
 use HrxDeliveryWoo\Debug;
 
 class Cronjob
@@ -15,13 +16,13 @@ class Cronjob
     public $cronjobs = array(
         'hrx_update_warehouses' => array(
             'func' => 'job_update_warehouses',
-            'freq' => 'daily',
+            'freq' => 'weekly',
             'time' => '02:00:00',
         ),
         'hrx_update_delivery_locations' => array(
             'func' => 'job_update_delivery_locs',
-            'freq' => 'monthly',
-            'time' => '03:00:00',
+            'freq' => 'weekly',
+            'time' => '04:00:00',
         ),
         /*'hrx_test' => array( // Activate if you want to test (use function job_test)
             'func' => 'job_test',
@@ -121,12 +122,19 @@ class Cronjob
 
     public function job_update_delivery_locs()
     {
-        $status = Terminal::update_delivery_locations();
-
+        $status = LocationsDelivery::update_couriers();
         if ( $status['status'] == 'OK' ) {
-            $debug_msg = 'Successfully updated delivery locations. Added ' . $status['added'] . ', updated ' . $status['updated'] . ', failed ' . $status['failed'];
+            $debug_msg = 'Successfully updated courier delivery locations. Added ' . $status['added'] . ', updated ' . $status['updated'] . ', failed ' . $status['failed'];
         } else {
-            $debug_msg = 'Failed to update delivery locations. Error: ' . $status['msg'];
+            $debug_msg = 'Failed to update courier delivery locations. Error: ' . $status['msg'];
+        }
+        Debug::to_log($debug_msg, 'cronjob', true);
+
+        $status = LocationsDelivery::update(false);
+        if ( $status['status'] == 'OK' ) {
+            $debug_msg = 'Successfully updated terminal delivery locations. Added ' . $status['added'] . ', updated ' . $status['updated'] . ', failed ' . $status['failed'];
+        } else {
+            $debug_msg = 'Failed to update terminal delivery locations. Error: ' . $status['msg'];
         }
         Debug::to_log($debug_msg, 'cronjob', true);
     }