+++ hrx-delivery-woo/classes/ShippingMethod.php
@@ -10,6 +10,7 @@ use HrxDeliveryWoo\Helper;
 use HrxDeliveryWoo\Core;
 use HrxDeliveryWoo\ShippingMethodHelper as ShipHelper;
 use HrxDeliveryWoo\ShippingMethodHtml as Html;
+use HrxDeliveryWoo\LocationsDelivery;
 use HrxDeliveryWoo\Debug;
 
 if ( ! class_exists('\HrxDeliveryWoo\ShippingMethod') ) {
@@ -84,6 +85,8 @@ if ( ! class_exists('\HrxDeliveryWoo\ShippingMethod') ) {
 
         public function init_form_fields()
         {
+            $all_wc_order_status = array_merge(array('' => '- '. __('Do not change', 'hrx-delivery') . ' -'), \wc_get_order_statuses());
+
             $fields = array(
                 'enable' => array(
                     'title' => __('Enable', 'hrx-delivery'),
@@ -161,7 +164,7 @@ if ( ! class_exists('\HrxDeliveryWoo\ShippingMethod') ) {
                 'id' => 'upd_delivery_loc',
                 'message' => __('Last update', 'hrx-delivery') . ':',
                 'default' => __('Never', 'hrx-delivery'),
-                'value' => Helper::get_hrx_option('last_sync_delivery_loc', ''),
+                'value' => Helper::get_hrx_option(LocationsDelivery::get_option_name('last_sync_terminal'), ''),
                 'repeat' => '30 days',
             );
 
@@ -186,6 +189,14 @@ if ( ! class_exists('\HrxDeliveryWoo\ShippingMethod') ) {
                     'custom_attributes' => array(
                         'data-key' => $method_key,
                     ),
+                );
+
+                $fields[$method_key . '_title'] = array(
+                    'title' => __('Method title', 'hrx-delivery'),
+                    'type' => 'select',
+                    'description' => __('Select the name of this shipping method that will be displayed on the Checkout page', 'hrx-delivery'),
+                    'options' => $this->get_shipping_method_title($method_params),
+                    'default' => 'sort',
                 );
 
                 $fields[$method_key . '_default_dimensions'] = array(
@@ -214,10 +225,32 @@ if ( ! class_exists('\HrxDeliveryWoo\ShippingMethod') ) {
                     'type' => 'prices_for_countries',
                     'countries' => $method_params['countries'],
                     'top_class' => 'hrx-method-' . $method_key,
+                    'empty_msg' => sprintf(__('The list of delivery locations was not received. Press the "%1$s" button at the "%2$s" parameter and reload the page after the update is complete (%3$s).', 'hrx-delivery'), __('Update locations', 'hrx-delivery'), '<a href="#check_token_button">' . __('Delivery locations', 'hrx-delivery') . '</a>', '<a href="javascript:location.reload();">' . __('Reload now', 'hrx-delivery') . '</a>'),
                 );
                 
                 $first = false;
             }
+
+            $fields['hr_wc_order_status'] = array(
+                'type' => 'hr',
+                'title' => __('Woocommerce Order status', 'hrx-delivery')
+            );
+
+            $fields['wc_status_on_ready'] = array(
+                'title' => __('WC order status when mark "Ready"', 'hrx-delivery'),
+                'type' => 'select',
+                'description' => __('Change WC order status to this, when HRX order is mark as "Ready"', 'hrx-delivery'),
+                'options' => $all_wc_order_status,
+                'default' => '',
+            );
+
+            $fields['wc_status_off_ready'] = array(
+                'title' => __('WC order status when unmark "Ready"', 'hrx-delivery'),
+                'type' => 'select',
+                'description' => __('Change WC order status to this, when HRX order is unmark as "Ready"', 'hrx-delivery'),
+                'options' => $all_wc_order_status,
+                'default' => '',
+            );
 
             /*$fields['hr_settings'] = array( //TODO: Disabled because section is empty
                 'type' => 'hr',
@@ -314,6 +347,17 @@ if ( ! class_exists('\HrxDeliveryWoo\ShippingMethod') ) {
             return Html::build_debug_plugin($value);
         }
 
+        private function get_shipping_method_title( $method, $get_title = false )
+        {
+            $all_titles = array(
+                'full' => $this->core->title . ' ' . $method['title'],
+                'sort' => 'HRX ' . strtolower($method['title']),
+                'method' => $method['title'],
+            );
+
+            return ($get_title && isset($all_titles[$get_title])) ? $all_titles[$get_title] : $all_titles;
+        }
+
         public function calculate_shipping($package = array())
         {
             $country = $package['destination']['country'];
@@ -355,9 +399,11 @@ if ( ! class_exists('\HrxDeliveryWoo\ShippingMethod') ) {
                 }
 
                 /* Build rate */
+                $title_key = (! empty($this->settings[$method_key . '_title'])) ? $this->settings[$method_key . '_title'] : 'sort';
+                $rate_title = $this->get_shipping_method_title($method_params, $title_key);
                 $rate = array(
                     'id' => ShipHelper::get_rate_id($method_key),
-                    'label' => $method_params['front_title'],
+                    'label' => $rate_title,
                     'cost' => $current_price,
                 );